Separate YAML formatting into its own plugin
This commit separates the yaml reporting code currently under Tern's
core code and puts it in its own plugin that Stevedore can utilize.
Specifically, the following changes were made:

1) Add yaml entrypoint to setup.cfg.

2) Remove the -y and --yaml command line option in tern/__main__.py.
   Moving forward we will utilize Stevedore to manage the plugin
   selection at runtime. Similar to the spdxtagvalue plugin, users can
   invoke the json reporting format by using "-m yaml" on the
   command line.

3) Copy the body of generate_yaml() in tern/report/report.py and move
   it under generate() in tern/formats/json/generator.py. Also copy
   print_yaml_report() function to tern/formats/json/generator.py so
   all the yaml-related formatting code is contained within the plugin.

4) Remove code corresponding to the -y command line option in
   tern/report/report.py.

5) Remove yaml_file constant in tern/utils/constants.py as it is no
   longer used.

6) Update the circleci test to run the yaml reporting using
   "-m yaml" instead of the deprecated -y command line option.

Resolves #370
