+++ ci/test_files_touched.py
@@ -76,7 +76,7 @@ test_suite = {
     # tern/report
     re.compile('tern/report'): [
         'tern -l report -i golang:alpine',
-        'tern -l report -y -i photon:3.0',
+        'tern -l report -m yaml -i photon:3.0',
         'tern -l report -s -i photon:3.0',
         'tern -l report -m json -i photon:3.0',
         'tern -l report -m spdxtagvalue -i photon:3.0',
