+++ setup.cfg
@@ -45,6 +45,7 @@ include_package_data = True
 tern.formats =
     spdxtagvalue = tern.formats.spdx.spdxtagvalue.generator:SpdxTagValue
     json = tern.formats.json.generator:JSON
+    yaml = tern.formats.yaml.generator:YAML
 console_scripts = 
     tern = tern.__main__:main
