+++ tern/__main__.py
@@ -121,9 +121,7 @@ def main():
                                metavar='REPORT_MODULE',
                                help="Format the report using one of the "
                                "available formats: "
-                               "spdxtagvalue, json")
-    parser_report.add_argument('-y', '--yaml', action='store_true',
-                               help="Create a report in yaml format")
+                               "spdxtagvalue, json, yaml")
     parser_report.add_argument('-f', '--file', default=None,
                                help="Write the report to a file; "
                                "If no file is given the default file in "
