+++ tern/report/content.py
@@ -7,8 +7,6 @@
 Functions to generate content for the report
 """
 
-import yaml
-
 from tern.command_lib import command_lib
 from tern.report import formats
 from tern.utils.general import get_git_rev_or_version
@@ -131,10 +129,3 @@ def print_summary_report(image):
                 notes = notes + print_package(package, '')
             notes = notes + formats.package_demarkation
     return notes
-
-
-def print_yaml_report(image):
-    '''Given an image object, create a yaml report'''
-    image_dict = {}
-    image_dict.update({'image': image.to_dict()})
-    return yaml.dump(image_dict, default_flow_style=False)
